requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Pillow==10.4.0
reportlab==4.2.2
gunicorn==21.2.0